(#1). Starting to add user configuration options
Adding option so that superupdate will just print a message prompting the user
(warn) instead of actually updating plugins. Also adding option for (interval)
and (days) which are the number of days between updates and the days when
updates are allowed/should be done.

Also starting adding these to the documentation, list variables require Vim7.