Add option to skip first check
This adds a configuration option 'g:superupdate_skip_first' that will
cause superupdate to skip trying to update plugins on it's first run.

This is useful in the case that superupdate is installed as part of a
batch (for instance when setting up a vim configuration on a new
machine) and all plugins are known to be up to date.

The option defaults to 'off', but when turned on during the first run it
will save the current date (as if it had updated) without actually
running the update.