Add support for vim-plug
Progress on: https://github.com/bradford-smith94/vim-superupdate/issues/2

This adds built-in support for the junegunn/vim-plug plugin manager. It
uses `globpath()` to detect if vundle or vim-plug are present in the
runtimepath. It also adds an autocmd to detect when vim-plug has been
called by the user manually in order to save the last update timestamp.

A potential bug could exist if a user manages plugins with vim-plug but
also has Vundle installed as superupdate will detect vundle first. For
now this is a wontfix because users *probably will not* have both
installed at the same time.