Better detect manual update commands
The previous autocmds only checked when closing update windows and could
have returned a false positive after running either `:PlugStatus` or
`:PluginList`. Both of those commands for vim-plug and vundle
respectively open a new window but do not actually update plugins.

This is solved by hooking an autocmd on the BufWinEnter event and
checking if the most recent command (in the '":' register) was an update
command.