Add support for neobundle and dein
Resolves https://github.com/bradford-smith94/vim-superupdate/issues/2

This adds (untested) support for the NeoBundle and Dein plugin managers.

This makes for a pretty good list of supported plugin managers and I'm
going to resolve the corresponding issue.

Also in this update are some fixes for linting warnings. Changing double
quoted strings to single quoted and adding 'abort' to functions in
autoload.