Fix issue updating using vim-plug in git directory
I have experienced an issue a few times where Vim is opened as my Git
commit editor and SuperUpdate fails to update my plugins using Vim-Plug.
This is a result of a fix to a Vim-Plug issue[1] where Vim-Plug throws
an exception[2] when certain Git environment variables are detected.

To fix this I have wrapped the actual update call to Vim-Plug in a
try/catch construct to gracefully fail in SuperUpdate.

[1]: https://github.com/junegunn/vim-plug/issues/506
[2]: https://github.com/junegunn/vim-plug/commit/5695fb84743ace811c03a5cd39ca7a825fa9caf7